Selfhost: Remove FASM, generate executable directly
Previously, the compiler would generate some assembly code, then would
hand it off to the fasm assembler to do the actual ELF creation. Now,
the ELF file is generated directly through the barely compiler, removing
fasm from the toolchain.

The main motivation for this was to achieve true independence. Without
this requirement on fasm, anyone with a 64 bit linux system, a copy of
the source code, and the compiler executable would be able to use
barely.

PS: I need to get better at incremental commits / changes...
Probably should have made a separate branch for this to then merge...