Compiler: Implement Scoping
Previously it was possible to reference a variable declared in some kind
of previous flow control, which could lead to unwanted undefined
behavior. Now only variables that are in the current scope or above will
be accessable.

Additionally, you can create a variable in an inner scope with the same
name as an already defined variable in a parent scop, use / modify the
variable, and it will not affect the parent scope variable.