Compiler: Remove auto-generated ZeroNode
Previously a variable that was not assigned to anything would be
automatically zero initialized. This behavior may not always be desired
and should be up to the user. The user can now choose to call the zero()
function which will return the zero initialized type.

Note: zero() will only work for direct assignment