Compiler: Force statements to not push onto stack
Previously, one was able to push a value onto the stack by simply
writing a statement that would not do anything with a value pushed onto
the stack. This is behavior that can be quite difficult to debug if
a function for example pushes data onto the stack that you are unaware
of. Now this code will throw a hard error at the specific line.

For the cases where pushing a value onto the stack is useful (ex: if
statement that is being used like an expression), ommitting the ';' at
the end of the line removes this check (similar to rust).

This concept happens to also work with returns (making it possible to
return an expression essentially) because of how a return
node is automatically added to the end of a function if one is not
specified in the code (also similar to rust).

One of the other reasons I am inclined to implement this is consistency,
'if' and 'while' statements don't have any sort of requirement for a
trailing semicolon after the expression they are evaluating, so I see no
reason one should be required in other cases where an expression is
being evaluated to be used in a larger context.