Compiler: Semi implemented function pointers
Works currently only if you store the pointer in a variable / use it as
a function argument, and then use that name to reference it before the
'('. This limitation will hopefully be pretty temporary, and is only a
result of difficulty handling situation directly calling a function
returned from another function.

Also the syntax for a function pointer type is not finalized, just the
most basic thing I could get working.