Compiler: Improve parsing of function pointer stuff
Instead of copying and saving the contents added by the parse_expression
before reloading it now parse_expression adds to a "fake" node list
instead of the main one. Then the data is copied from the fake to the
real once needed.

This change is also slightly better (at least in my opinion), aka less
weird. But most importantly it takes about 1/2 the time to recompile the
compiler compared to the previous implementation.