Compiler: Remove "include"
I do not like the idea of include anymore, and I think that it would get
in the way of a better import system. Include is not really suitable for
projects outside the actual barely development directory.

I want to treat the barely compiler just as a executable that takes in
input files and outputs a binary, as opposed to one that does a bunch of
other heavy lifting.

The idea to eventually make this system not tedious (having the manually
include all the files) is to create a separate build tool which will
collect all the sources and then run the actual barely compiler with
them.