Compiler: Parse if & while statement bodies as expressions
Now curly braces are technically not required, which could be useful in
some scenaries where there are really simple bodies.

Even ignoring the actual new syntax added, this is probably a pretty
good idea because it helps reuse code. Now the inside of an if or while
statement is just parsed like an expression, and in most cases that
results in it being parsed like a block, so now there aren't two places
I am parsing things like blocks.