Implement better getter and setter syntax
Syntax not set in stone yet, however considering the fact that the
majority of modern languages use this syntax, it probably isn't too bad
of an idea.

The new syntax will make it much easier to write and understand code
written in barely, as working with structures would now be more of a
first class feature on the language level.