New way of casting
I currently want to limit there to be only one way of doing things, and
the old casting sometimes required extra parenthesis, so the new way
will be the only way to do things for now.

Potentionally in the future the old way will be readded, so that they
both are options.

New: cast(Type, expr)
Old: cast(Type) expr